// src/project/exceptions/project.enum.ts
export enum ProjectError {
  NOT_FOUND = 'Project not found',
  ALREADY_EXISTS = 'Project already exists',
  SOMETHING_WENT_WRONG_CREATING = 'Something went wrong while creating a project',
  SOMETHING_WENT_WRONG_FETCHING = 'Something went wrong while fetching projects',
  SOMETHING_WENT_WRONG_FETCHING_ONE = 'Something went wrong while fetching a project',
  SOMETHING_WENT_WRONG_DELETING = 'Something went wrong while deleting a project',
  SOMETHING_WENT_WRONG_UPDATING = 'Something went wrong while updating a project',
}